refactor(websocket): reorganize document handling and remove unused modules
- Removed the `doc` module and its associated handler, streamlining the codebase.
- Introduced a new `http` module under `domain::value_objects` to encapsulate document-related data structures.
- Updated routing and handler references to align with the new structure, ensuring proper functionality for document operations.
