--- server/websocket/src/doc/mod.rs
@@ -1,7 +1,0 @@
-pub mod handler;
-pub mod route;
-mod types;
-
-pub use handler::DocumentHandler;
-pub use route::document_routes;
-pub use types::{Document, DocumentResponse, HistoryItem, HistoryResponse, RollbackRequest};
