+++ server/websocket/src/domain/value_objects/mod.rs
@@ -1,2 +1,3 @@
 pub mod conf;
+pub mod http;
 pub mod redis;
